requirements.txt
streamlit>=1.16
tensorflow>=2.8
numpy
pillow